# backend/docs/swagger.yaml
definitions:
  gin_http.ResponseJSON:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/search/user:
    get:
      parameters:
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 搜索用户
...
  /api/v1/user/{id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 获取用户信息
  /api/v1/user/{id}/articles:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 用户动态列表
  /api/v1/user/{id}/avatar:
    put:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: image
        in: formData
        name: image
        required: true
        type: file
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 更新用户头像
  /api/v1/user/{id}/fans:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 粉丝列表
  /api/v1/user/{id}/follow:
    post:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 关注用户ID
        in: formData
        name: follow_id
        required: true
        type: integer
      - description: 类型
        in: formData
        name: type
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 关注用户
  /api/v1/user/{id}/follows:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 关注列表
  /api/v1/user/{id}/likeArticles:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: page_num
        in: query
        name: page_num
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 用户喜欢列表
  /user/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin_http.ResponseJSON'
      summary: 注销用户
...
swagger: "2.0"